pyproject.toml
[project]
name = "kagi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.4",
    "robyn>=0.64.1",
    "scrapy>=2.12.0",
    "sqlean-py>=3.47.0",
    "tantivy>=0.22.0",
    "trafilatura>=1.12.2",
]
